refactor(spec): split schema facade, normalize-input, capabilities, tier2 tests
Pure mechanical splits, no behavior change, export surfaces byte-identical:

- schema.ts (573 -> 283) -> schema-imports.ts (SpecDefsRoot/reRootSpec/
  SpecModule + 66 Import constants stay together) + schema-types.ts
  (Static type bag incl. hand-written DataRef/BlankParams overrides);
  schema.ts is a pure re-export facade, 113 import sites untouched;
  schema/v0.json byte-identical
- normalize-input.ts (574 -> 154) -> layers-marks (29 families) +
  layers-special (20 geoms + 49-arm union) behind re-export facade
- capabilities.ts (522 -> 159) -> capabilities-data.ts (SCALE_CAPABILITIES
  leaf) + capabilities-helpers.ts; cycle-free (data is a leaf)
- validate-tier2-color.test.ts (862 -> 182) + censor-a/b files (308/392)
- validate-tier2-data.test.ts (788 -> 333) + -breaks (365) + -profile (142)

Expanded test-name multisets identical (47 + 24 tests). Ratchet pruned
56 -> 51. Gates: bun test packages/spec 771 pass, bun run check, knip,
lint:type-aware:run, schema:emit diff, max-loc --all all clean.
