+++ scripts/guards/max-loc-exclusions.txt
@@ -20,11 +20,9 @@ packages/svelte/tests/inspection/resolve-snapshot.test.ts 940 # legacy oversize;
 packages/svelte/src/lib/data/kyoto-sakura.ts 891 # pure data corpus, no logic — permanently exempt
 packages/svelte/src/lib/plot-engine.svelte.ts 883 # legacy oversize; pending refactor slice S5
 scripts/sakura-lesson.test.ts 870 # legacy oversize; pending refactor slice S7
-packages/spec/tests/validate-tier2-color.test.ts 862 # legacy oversize; pending refactor slice S4
 apps/docs/src/lib/theme-specimens/data.ts 821 # pure data corpus, no logic — permanently exempt
 examples/point/stat-unique/data.ts 796 # pure data corpus, no logic — permanently exempt
 examples/line/ecdf/data.ts 793 # pure data corpus, no logic — permanently exempt
-packages/spec/tests/validate-tier2-data.test.ts 788 # legacy oversize; pending refactor slice S4
 packages/svelte/tests/inspection/coordinator.test.ts 736 # legacy oversize; pending refactor slice S6
 packages/svelte/src/lib/surface/surface-state.svelte.ts 717 # legacy oversize; pending refactor slice S5
 scripts/quickstart/steps.ts 698 # legacy oversize; pending refactor slice S7
@@ -42,8 +40,6 @@ packages/svelte/src/lib/interval/interval-state.svelte.ts 583 # legacy oversize;
 scripts/gen-benchmark-charts.ts 583 # legacy oversize; pending refactor slice S7
 packages/svelte/tests/legend/focus-entry-key-index-pure.test.ts 581 # legacy oversize; pending refactor slice S6
 packages/svelte/tests/legend/focus-interaction.test.ts 577 # legacy oversize; pending refactor slice S6
-packages/spec/src/normalize-input.ts 574 # legacy oversize; pending refactor slice S4
-packages/spec/src/schema.ts 572 # legacy oversize; pending refactor slice S4
 packages/svelte/src/lib/GGPlot.svelte 570 # legacy oversize; pending refactor slice S5
 scripts/skill-content.test.ts 563 # legacy oversize; pending refactor slice S7
 packages/svelte/tests/inspection/inspection-state.pin-queue.test.ts 544 # legacy oversize; pending refactor slice S6
@@ -52,7 +48,6 @@ packages/svelte/tests/runtime/semantic-candidate-projection.test.ts 543 # legacy
 examples/point/gradient-continuous/data.ts 530 # pure data corpus, no logic — permanently exempt
 packages/svelte/src/lib/scene/Batch.svelte 526 # legacy oversize; pending refactor slice S5
 benchmarks/competitive/fixtures/main.ts 523 # legacy oversize; pending refactor slice S7
-packages/spec/src/capabilities.ts 522 # legacy oversize; pending refactor slice S4
 packages/svelte/tests/selection/selection.test.ts 521 # legacy oversize; pending refactor slice S6
 scripts/gen-llms.test.ts 521 # legacy oversize; pending refactor slice S7
 packages/svelte/tests/inspection/inspection-teardown.test.ts 518 # legacy oversize; pending refactor slice S6
